requirements.txt
# General Requirements
nose==1.2.1
chardet==2.1.1
iso8601==0.1.4

# App Requirements
PIL==1.1.7
django==1.4.5
south==0.7.6
django-guardian==1.0.4
django-debug-toolbar==0.9.4
django-extensions==1.0.2
django-jsonfield==0.8.12
gargoyle==0.10.6
pyyaml==3.10
django-ratings==0.3.7
django-nose==1.1
django-storages==1.1.5
django_compressor==1.1.2
BeautifulSoup<=4.0
html5lib==0.95
slimit==0.7.4
cssmin==0.1.4
django-analytical==0.14

